# src/main/java/pl/wsb/students/model/ActorList.java
/*
 * Java WSB Course Sample Movie App API
 * Sample Movie App API
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package pl.wsb.students.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.util.ArrayList;
import java.util.List;
import javax.validation.constraints.*;

/**
 * ActorList
 */
@javax.annotation.Generated(value = "io.swagger.codegen.languages.JavaJerseyServerCodegen", date = "2020-04-24T11:57:33.468Z")
public class ActorList   {
  @JsonProperty("total")
  private Integer total = null;

  @JsonProperty("data")
  private List<Actor> data = new ArrayList<Actor>();

  public ActorList total(Integer total) {
    this.total = total;
    return this;
  }

  /**
   * Actor list total items count
   * @return total
   **/
  @JsonProperty("total")
  @ApiModelProperty(required = true, value = "Actor list total items count")
  @NotNull
  public Integer getTotal() {
    return total;
  }

  public void setTotal(Integer total) {
    this.total = total;
  }

  public ActorList data(List<Actor> data) {
    this.data = data;
    return this;
  }

  public ActorList addDataItem(Actor dataItem) {
    this.data.add(dataItem);
    return this;
  }

  /**
   * One pagination page data
   * @return data
   **/
  @JsonProperty("data")
  @ApiModelProperty(required = true, value = "One pagination page data")
  @NotNull
  public List<Actor> getData() {
    return data;
  }

  public void setData(List<Actor> data) {
    this.data = data;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    ActorList actorList = (ActorList) o;
    return Objects.equals(this.total, actorList.total) &&
        Objects.equals(this.data, actorList.data);
  }

  @Override
  public int hashCode() {
    return Objects.hash(total, data);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class ActorList {\n");
    
    sb.append("    total: ").append(toIndentedString(total)).append("\n");
    sb.append("    data: ").append(toIndentedString(data)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}